Dealing with strange mismatch in block length

+++ src/processing/processing_imagesplit.hpp
@@ -30,9 +30,9 @@ namespace sgns::processing
         */
         ImageSplitter(
             const char* filename,
-            uint32_t blockstride,
-            uint32_t blocklinestride,
-            uint32_t blocklen
+            uint64_t blockstride,
+            uint64_t blocklinestride,
+            uint64_t blocklen
         );
 
         /** Split an image loaded from raw data of a file loaded elsewhere, i.e. asynciomanager
@@ -42,9 +42,9 @@ namespace sgns::processing
         * @param blocklen - Block Length in bytes
         */
         ImageSplitter(const std::vector<char>& buffer,
-            uint32_t blockstride,
-            uint32_t blocklinestride,
-            uint32_t blocklen);
+            uint64_t blockstride,
+            uint64_t blocklinestride,
+            uint64_t blocklen);
         
         /** Split an image loaded from raw RGBA bytes
         * @param buffer - Raw RGBA
@@ -53,9 +53,9 @@ namespace sgns::processing
         * @param blocklen - Block Length in bytes
         */
         ImageSplitter(const std::vector<uint8_t>& buffer,
-            uint32_t blockstride,
-            uint32_t blocklinestride,
-            uint32_t blocklen);
+            uint64_t blockstride,
+            uint64_t blocklinestride,
+            uint64_t blocklen);
 
         ~ImageSplitter()
         {
@@ -110,9 +110,9 @@ namespace sgns::processing
         std::vector<std::vector<uint8_t>> splitparts_;
         int partwidth_ = 32;
         int partheight_ = 32;
-        uint32_t blockstride_;
-        uint32_t blocklinestride_;
-        uint32_t blocklen_;
+        uint64_t blockstride_;
+        uint64_t blocklinestride_;
+        uint64_t blocklen_;
         const unsigned char* inputImage;
         size_t imageSize;
         std::vector<int> chunkWidthActual_;
